pyproject.toml
[tool.poetry]
name = "fastapi2"
version = "0.1.0"
description = ""
authors = ["anvtar <annatrn03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.1"
psycopg = "^3.1.16"
asyncpg = "^0.29.0"
postgres = "^4.0"
sqlalchemy = "^2.0.25"
fastapi-sqlalchemy = "^0.2.1"
pydantic-settings = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
